refactor: 71 removed and rearrange files
