--- 71
@@ -1,2 +1,0 @@
-'taskill'은(는) 내부 또는 외부 명령, 실행할 수 있는 프로그램, 또는
-배치 파일이 아닙니다.
